+++ gulpfile.js
@@ -1,3 +1,3 @@
 var gulp = require('gulp'),
-  options = require('./tasks/sass-options.js');
+  options = require('./task-sets/sass-options.js');
 require('sl-gt-sass-autoprefixer')(gulp, options);
